Transform project into Maven project to more easily handle dependencies and building
