--- src/TF2ClassWarsStatTracker/util/Constants.java
@@ -1,14 +1,0 @@
-package TF2ClassWarsStatTracker.util;
-
-import java.awt.*;
-
-public class Constants {
-    public static final int BLU = 0, RED = 1;
-    public static final Color
-            BLU_COLOUR = new Color(171,203,255),
-            RED_COLOUR = new Color(255,125,125);
-    public static final String[] MERCENARY = {
-            "Scout", "Soldier", "Pyro",
-            "Demoman", "Heavy", "Engineer",
-            "Medic", "Sniper", "Spy"};
-}
